move api types to their respective packages
