--- packages/libs/api/package.json
@@ -1,14 +1,0 @@
-{
-  "private": true,
-  "name": "@truckermudgeon/api",
-  "version": "0.0.0",
-  "license": "GPL-3.0-or-later",
-  "type": "module",
-  "sideEffects": false,
-  "exports": {
-    "./*": "./*.ts"
-  },
-  "devDependencies": {
-    "typescript": "^5.4.2"
-  }
-}
